# Optimize creation and verification of transactions with lots of inputs