--- a/src/wallet/wallet.cpp
+++ b/src/wallet/wallet.cpp
@@ -1996,8 +1996,19 @@ bool CWallet::CreateTransaction(const vector<CRecipient>& vecSend, CWalletTx& wt
                     bool signSuccess;
                     const CScript& scriptPubKey = coin.first->vout[coin.second].scriptPubKey;
                     CScript& scriptSigRes = txNew.vin[nIn].scriptSig;
+                    // Sign the first input with SIGHASH_ALL, all the rest with
+                    // SIGHASH_ALL|SIGHASH_ANYONECANPAY. This optimizes validation for
+                    // transactions with a lot of inputs. The first SIGHASH_ALL prevents
+                    // somebody from adding additional inputs and changing the transaction id;
+                    // SIGHASH_ALL by itself fixes the set of previous outputs.
+                    // Using SIGHASH_ANYONECANPAY for the rest of the inputs minimizes the
+                    // amount of data that has to be hashed to compute signatures, making both
+                    // transaction creation and verification faster (benchmark results;
+                    // a 500 input transaction takes 277ms to validate with this optimization,
+                    // but 342ms if all inputs are SIGHASH_ALL).
+                    int sighash_type = (nIn == 0 ? SIGHASH_ALL : SIGHASH_ALL|SIGHASH_ANYONECANPAY);
                     if (sign)
-                        signSuccess = ProduceSignature(TransactionSignatureCreator(this, &txNewConst, nIn, SIGHASH_ALL), scriptPubKey, scriptSigRes);
+                        signSuccess = ProduceSignature(TransactionSignatureCreator(this, &txNewConst, nIn, sighash_type), scriptPubKey, scriptSigRes);
                     else
                         signSuccess = ProduceSignature(DummySignatureCreator(this), scriptPubKey, scriptSigRes);