@gavinandresen 2015-07-28
Self-reviewing after having a chance to sleep on this:

This change would mean an unconfirmed transaction gives potential attackers some inputs that belong to me, but are signed so that they are committed to sending the coins to a particular set of outputs but NOT to-a-particular-set-of-outputs-but-only-when-bundled-with-a-particular-set-of-inputs.

So an attacker could:
1) Replace the SIGHASH_ALL first input with one of their inputs. Costs the attacker however many coins are in the first input, not a very interesting attack (unless part of a larger attack that requires transaction malleability to pull off).

2) If the first SIGHASH_ALL input's amount is less than the transaction fee, attacker could relay/broadcast without that first input. Not very interesting, worst case the alternate
lower-fee transaction gets mined, user saves some coin.

3) If the wallet has TWO or more unconfirmed, multi-input transactions with exactly the same outputs (paying exactly same amount to same address-- if change is generated, then  change address/amount exactly the same for both), then there is a worrisome attack:  attacker can combine all of the SIGHASH_ANYONECANPAY inputs from the transactions into one big transaction that has an oversize fee, and mine that instead of the two transactions.

We can say "don't re-use addresses" ... but (3) is a good reason not to merge this as-is