@gavinandresen 2015-08-13
I'm going to do a brain-dump on what it would take to make this change safe... but close this pull request. The performance improvement for transactions with lots of inputs is significant, but those transactions are pretty rare.

So, to do this safely, the wallet should only sign with SIGHASH_ANYONECANPAY if it knows it has never signed any other transactions with an identical set of outputs.

The naive way of doing that would be to iterate over all previous wallet 'send' transactions and check.

Or the code could keep track of whether or not a brand-new change address was being used, and only use ANYONECANPAY if it was.

Or the wallet could keep a bloom filter-- for each 'send' transaction in the wallet, add a hash of just the transaction's outputs to the filter. When sending a new transaction, if it's outputs don't match the bloom filter, it is safe to use ANYONECANPAY. If it does match the bloom filter, then assume it is not safe (it might be a false-positive, and that's OK).

It is not clear to me whether actually implementing that code is worth it; most wallets probably don't create transactions with lots of inputs, and maintaining the bloom filter is extra work that isn't needed in most cases.