@dcousens 2015-08-13
This would also make transactions highly identifiable as coming from `bitcoin-qt`,  as this is a publicly visible (e.g on the blockchain) implementation detail.