@sipa 2015-08-13
They're already identifiable to a significant extent (fee sniping, earlier
due to randomization of outouts/change and non-reuse of change addresses).
I din't think that's a worry. We should encourage best practices.

@gavinandresen Considered adding an OP_RETURN with a bit of random data to
make the outputs unique, for very large transactions?