@dcousens 2015-08-13
> They're already identifiable to a significant extent (fee sniping, earlier
due to randomization of outouts/change and non-reuse of change addresses).

Fee sniping is on its [last legs](https://github.com/bitcoin/bitcoin/pull/6216) though right?

> We should encourage best practices.

Of course,  but would encouraging this for the sake of performance in this client,  be considered best practice for other implementations?