@sipa 2015-08-13
Not only performance of the client. Performance of validation by others,
which for very large transactions may be in thw best interest of the one
creating it.