@gavinandresen 2015-08-13
@sipa : ooh, I like the add-an-extra-OP_RETURN idea....